fix: `make clean` freeze sometime when install

+++ scripts/build_athrill.sh
@@ -107,13 +107,15 @@ if [ "$1" = "check" ]; then
 fi
 
 if [ -z "$CHECK" ]; then
-    cd "$ETROBO_ATHRIL_WORKSPACE" > /dev/null 2>&1
-    echo make ASP3 workspace clean
-    make clean > /dev/null 2>&1
-    rm -f asp
-    cd "$ETROBO_ATHRILL_TARGET" > /dev/null 2>&1
-    echo make Athrill clean
-    make timer32=true clean > /dev/null 2>&1
+    if [ -d "$ETROBO_ATHRIL_WORKSPACE" ]; then
+        cd "$ETROBO_ATHRIL_WORKSPACE"
+        echo make ASP3 workspace clean
+        make clean > /dev/null 2>&1
+        rm -f asp
+        cd "$ETROBO_ATHRILL_TARGET"
+        echo make Athrill clean
+        make timer32=true clean > /dev/null 2>&1
+    fi
 fi
 
 cd "$ETROBO_ROOT"
